refactor: Reservation DTO 패키지 분리
